Share main menu markup and styling between homepage and rest of site
Since the styles for the homepage are completely separate
from the styles for the rest of the site, the shared menu
required some slightly unintuitive SCSS @imports.

But it’s still more sensible than duplicating all the
menu styling, once for the homepage and once for the rest
of the site.

Fixes #34.